#6273: move the subclass tests back into PhDefaultTest, drop ObservantTest
